検査の陽性率: Rename components/cards/PositiveRateCard.vue to components/index/CardsMonitoring/PositiveRate/Card.vue
